_archive_2025_07_30/duplicate_workspaces/master_workspace/requirements.txt
# Master Workspace Python Dependencies
# ====================================

# Core framework
agno>=2.0.0

# Async libraries
asyncio
aiohttp>=3.8.0
asyncpg>=0.28.0

# Database connectors  
redis>=4.5.0
clickhouse-connect>=0.6.0
asyncio-clickhouse>=0.1.0

# Docker integration
docker>=6.0.0

# Monitoring and observability
prometheus-client>=0.16.0
opentelemetry-api>=1.17.0
opentelemetry-sdk>=1.17.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# Configuration
pyyaml>=6.0
pydantic>=2.0.0

# Utilities
python-dateutil>=2.8.0
psutil>=5.9.0
requests>=2.28.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Machine Learning (optional)
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0

# Web interface (Agno 標準)
streamlit>=1.28.0
fastapi>=0.95.0
uvicorn>=0.20.0
jinja2>=3.1.0